+++ samples/go.mod
@@ -17,7 +17,7 @@ require (
 	gocloud.dev/pubsub/natspubsub v0.15.0
 	gocloud.dev/pubsub/rabbitpubsub v0.15.0
 	gocloud.dev/runtimevar/etcdvar v0.15.0
-	gocloud.dev/secrets/vaultsecrets v0.15.0
+	gocloud.dev/secrets/hashivault v0.15.0
 	google.golang.org/genproto v0.0.0-20190605220351-eb0b1bdb6ae6
 	gopkg.in/pipe.v2 v2.0.0-20140414041502-3c2ca4d52544
 )
@@ -34,4 +34,4 @@ replace gocloud.dev/pubsub/rabbitpubsub => ../pubsub/rabbitpubsub
 
 replace gocloud.dev/runtimevar/etcdvar => ../runtimevar/etcdvar
 
-replace gocloud.dev/secrets/vaultsecrets => ../secrets/vaultsecrets
+replace gocloud.dev/secrets/hashivault => ../secrets/hashivault
