+++ samples/gocdk-secrets/main.go
@@ -32,7 +32,7 @@ import (
 	_ "gocloud.dev/secrets/azurekeyvault"
 	_ "gocloud.dev/secrets/gcpkms"
 	_ "gocloud.dev/secrets/localsecrets"
-	_ "gocloud.dev/secrets/vaultsecrets"
+	_ "gocloud.dev/secrets/hashivault"
 )
 
 const helpSuffix = `
